+++ BraintreeCard/BTAuthenticationInsight_Internal.h
@@ -1,0 +1,7 @@
+#import "BTAuthenticationInsight.h"
+
+@interface BTAuthenticationInsight ()
+
+- (instancetype)initWithJSON:(BTJSON *)json;
+
+@end