+++ BraintreeCard/BTCard_Internal.h
@@ -9,4 +9,6 @@
 
 extern NSString * const BTCardGraphQLTokenizationMutation;
 
+extern NSString * const BTCardGraphQLTokenizationWithAuthenticationInsightMutation;
+
 @end