+++ BraintreeCard/Public/BTCard.h
@@ -127,6 +127,16 @@ Designated initializer.
 */
 @property (nonatomic, assign) BOOL shouldValidate;
 
+/**
+ Optional: If authentication insight is requested. If this property is set to true, a `merchantAccountId` must be provided. Defaults to false.
+ */
+@property (nonatomic, assign) BOOL authenticationInsightRequested;
+
+/**
+ Optional: The merchant account id.
+ */
+@property (nonatomic, nullable, copy) NSString *merchantAccountId;
+
 @end
 
 NS_ASSUME_NONNULL_END